Created GitHub Actions, Pull Request Workflow. (#5)
* Created a gitignore files to omit unnecessary cache files.

* Created initial pytest file.

This file specifies the naming convention for all test files.
All test files from this point forward are required to follow the
following formula:
  test_<test_name>.py

* Created initial unit test to set up pytest within github workflows.

This unit test is a simple "Hello world!" unit test and can/will
be removed once more applicable code and unit tests are created.

* Added a Pull Request workflow for GitHub Actions.

The workflow has the following jobs and steps:
- Pylint
  - Set up python.
  - Install dependencies.
  - Lint python code using flake8.
- Pytest
  - Set up python.
  - Install dependencies.
  - Test unit tests and code using pytest.
- Codecov
  - Set up python.
  - Install dependencies.
  - Generate coverage report.
  - Upload coverage to Codecov.

* Remove token reference within Codecov job temporarily.

This token reference is not required for public repositories. However,
it may be added back afterwards as codecov will set up a proper token
afterwards.

* Added token reference with Codecov job.

The token from codecov was generate in between this commit and the
previous. It has been added back as the connection will now exist.